+++ packages/vendure-plugin-address-lookup/src/api/address-lookup.resolver.ts
@@ -1,6 +1,6 @@
 import { Args, Query, Resolver } from '@nestjs/graphql';
 import { OrderAddress } from '@vendure/common/lib/generated-types';
-import { Ctx, RequestContext } from '@vendure/core';
+import { Ctx, ForbiddenError, RequestContext } from '@vendure/core';
 import { QueryLookupAddressArgs } from '../generated/graphql';
 import { AddressLookupService } from '../services/address-lookup.service';
 
@@ -13,10 +13,10 @@ export class AddressLookupResolver {
     @Ctx() ctx: RequestContext,
     @Args() args: QueryLookupAddressArgs
   ): Promise<OrderAddress[]> {
-    // if (!ctx.session?.activeOrderId) {
-    //   // Prevent abuse, so only allow calls with an active order
-    //   throw new ForbiddenError();
-    // }
+    if (!ctx.session?.activeOrderId) {
+      // Prevent abuse, so only allow calls with an active order
+      throw new ForbiddenError();
+    }
     return await this.addressLookupService.lookupAddress(ctx, args.input);
   }
 }