+++ packages/vendure-plugin-promotion-extensions/src/constants.ts
@@ -1,0 +1,4 @@
+export const PROMOTION_EXTENSIONS_PLUGIN_OPTIONS = Symbol(
+  'PROMOTION_EXTENSIONS_PLUGIN_OPTIONS'
+);
+export const loggerCtx = 'PromotionExtensionsPlugin';